requirements.txt
fastapi
uvicorn[standard]
langgraph>=0.1.0
langchain-core>=0.1.0
google-generativeai>=0.5.0
redis[asyncio]>=5.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
plotly>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
prophet>=1.1.0
faker>=24.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
